Mejorando diseño de la vista de Mi cuenta agregando blloques para navegacion en espera de componentes faltantes
