+++ system_rentail/src/styles/my-account.sass
@@ -1,0 +1,125 @@
+@import './variables.sass'
+
+.box-account
+    width: 96%
+    padding: 2%
+    padding-top: 0%
+    display: flex
+    justify-content: center
+    align-items: center
+
+    .profile
+        width: 20%
+        margin: 3%
+        padding: 1%
+        background-color: #fff
+        box-shadow: 0.5px 0.5px 0.5px 0.5px #808080
+
+        figure
+            width: 100%
+            margin: 0 auto
+            margin-bottom: 5%
+            display: flex
+            align-items: center
+            justify-content: center
+
+        .data-profile
+            width: 100%
+            flex-direction: column
+            display: flex
+            justify-content: center
+            align-items: center
+
+            .option
+                width: 100%
+                display: flex
+                justify-content: center
+                align-items: center
+
+                .input-profile
+                    width: 100%
+                    margin: 3%
+
+        a
+            margin-top: 3%
+            cursor: pointer
+            color: #3468cf
+            font-size: 0.9em
+            &:hover
+                color: #153E90
+                font-weight: bold
+
+    .menu-account
+        width: 60%
+        margin: 3%
+        padding: 2%
+        background-color: #fff
+        box-shadow: 0.5px 0.5px 0.5px 0.5px #808080
+        display: grid
+        grid-template-columns: repeat(2, 1fr)
+        gap: 3%
+
+@media (max-width: $small)
+    .box-account
+        width: 98%
+        padding: 1%
+        flex-direction: column-reverse
+        .profile
+            width: 90%
+            margin: 2%
+            padding: 2%
+
+        .menu-account
+            width: 90%
+            padding: 2%
+            text-align: center
+
+            .text-card
+                font-size: 1em
+
+@media (max-width: $medium) and (min-width: $small + 1px)
+    .box-account
+        width: 98%
+        padding: 1%
+        flex-direction: column-reverse
+        .profile
+            width: 90%
+            margin: 2%
+            padding: 2%
+
+        .menu-account
+            width: 90%
+            padding: 2%
+
+            .card-media
+                height: 130px
+
+            .text-card
+                font-size: 1.5em
+
+@media (max-width: $medium-large) and (min-width: $medium + 1px)
+    .box-account
+        width: 98%
+        padding: 1%
+        .profile
+            width: 35%
+            margin: 2%
+            padding: 1%
+
+        .menu-account
+            width: 50%
+
+            .text-card
+                font-size: 1.5em
+
+@media (max-width: $large) and (min-width: $medium-large + 1px)
+    .box-account
+        width: 98%
+        padding: 1%
+        .profile
+            width: 35%
+            margin: 2%
+            padding: 1%
+
+        .menu-account
+            width: 55%
