--- system_rentail/src/styles/profile.sass
@@ -1,16 +1,0 @@
-.box-profile
-    width: 90%
-    padding: 5%
-    display: grid
-    grid-template-row: repeat(2, 1fr)
-    justify-content: center
-    align-items: center
-
-    .data-profile
-        width: 100%
-        margin: 0 auto
-        display: grid
-        grid-template-columns: repeat(1, 1fr)
-        justify-content: center
-        align-items: center
-        gap: 5%
